backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
